Refactor interactive HTML asset generation
