+++ ChartForgeX.Examples/GalleryWriter.cs
@@ -120,13 +120,7 @@ public static partial class GalleryWriter {
         sb.AppendLine("<meta name=\"viewport\" content=\"width=device-width, initial-scale=1\">");
         sb.AppendLine("<title>ChartForgeX SVG/PNG visual comparison</title>");
         sb.AppendLine("<style>");
-        sb.AppendLine(":root{color-scheme:dark;--bg:#0f172a;--panel:#111827;--frame:#020617;--line:#334155;--text:#e5e7eb;--muted:#94a3b8;--ok:#86efac;--warn:#fbbf24;--accent:#38bdf8}");
-        sb.AppendLine("*{box-sizing:border-box}body{margin:0;padding:24px;background:var(--bg);color:var(--text);font-family:Inter,ui-sans-serif,system-ui,Segoe UI,Arial,sans-serif}");
-        sb.AppendLine("header{max-width:1500px;margin:0 auto 24px}h1{margin:0 0 6px;font-size:24px;line-height:1.15}p{margin:0;color:var(--muted)}.summary,.family-nav{display:flex;flex-wrap:wrap;gap:8px;margin-top:14px}.pill,.family-nav a{border:1px solid var(--line);border-radius:999px;padding:6px 10px;color:#cbd5e1;font-size:12px;font-weight:700;text-decoration:none}.pill.ok,.family-nav a.ok{border-color:rgba(134,239,172,.45);color:var(--ok)}.pill.warn,.family-nav a.warn{border-color:rgba(251,191,36,.55);color:var(--warn)}a.pill:hover,.family-nav a:hover{border-color:rgba(56,189,248,.55);color:var(--accent)}main{display:grid;gap:34px;max-width:1500px;margin:0 auto}");
-        sb.AppendLine(".family{display:grid;gap:16px}.family-head{display:flex;align-items:end;justify-content:space-between;gap:16px;border-bottom:1px solid var(--line);padding-bottom:12px}.family-title{margin:0 0 5px;font-size:19px;line-height:1.15;color:#f8fafc}.family-desc{font-size:13px}.family-count{flex:0 0 auto;border:1px solid rgba(56,189,248,.42);border-radius:999px;color:var(--accent);font-size:12px;font-weight:800;padding:6px 10px}.family-count.ok{border-color:rgba(134,239,172,.45);color:var(--ok)}.family-count.warn{border-color:rgba(251,191,36,.55);color:var(--warn)}section{background:var(--panel);border:1px solid var(--line);border-radius:8px;padding:16px;scroll-margin-top:18px}section.mismatch{border-color:rgba(251,191,36,.6)}h2{display:flex;align-items:center;justify-content:space-between;gap:12px;margin:0 0 14px;font-size:14px;color:#cbd5e1}.status{border:1px solid var(--line);border-radius:999px;padding:4px 8px;font-size:11px;font-weight:800}.status.ok{color:var(--ok);border-color:rgba(134,239,172,.45)}.status.warn{color:var(--warn);border-color:rgba(251,191,36,.55)}.pair{display:grid;grid-template-columns:repeat(2,minmax(0,1fr));gap:16px;align-items:start}section.large .pair{grid-template-columns:1fr}.wipe-figure{grid-column:1/-1;order:-1}.wipe-figure .media{min-height:360px}section.large .wipe-figure .media{min-height:min(68vh,820px)}section.large figure:not(.wipe-figure) .media{min-height:min(62vh,720px)}");
-        sb.AppendLine("figure{margin:0;background:var(--frame);border:1px solid #1f2937;border-radius:8px;padding:10px;min-width:0}.caption{display:flex;justify-content:space-between;gap:12px;color:var(--muted);font-size:12px;margin-bottom:8px}.dims{font-variant-numeric:tabular-nums;color:#cbd5e1}.format{color:var(--accent);font-weight:800;text-decoration:none}.format:hover{text-decoration:underline}.wipe-controls{display:flex;flex-wrap:wrap;gap:6px;margin:0 0 10px}.wipe-controls input{position:absolute;opacity:0;pointer-events:none}.wipe-controls label{border:1px solid var(--line);border-radius:999px;color:#cbd5e1;cursor:pointer;font-size:11px;font-weight:800;padding:5px 9px}.wipe-controls input:focus-visible+label{outline:2px solid var(--accent);outline-offset:2px}.wipe-controls input:checked+label{background:rgba(56,189,248,.16);border-color:rgba(56,189,248,.7);color:#f8fafc}");
-        sb.AppendLine(".media{width:100%;min-height:140px;background-color:#020617;background-image:linear-gradient(45deg,#0b1120 25%,transparent 25%),linear-gradient(-45deg,#0b1120 25%,transparent 25%),linear-gradient(45deg,transparent 75%,#0b1120 75%),linear-gradient(-45deg,transparent 75%,#0b1120 75%);background-size:24px 24px;background-position:0 0,0 12px,12px -12px,-12px 0;overflow:hidden}.media.wipe{position:relative;--wipe:50%}.wipe-controls .wipe-25:checked~.media{--wipe:25%}.wipe-controls .wipe-50:checked~.media{--wipe:50%}.wipe-controls .wipe-75:checked~.media{--wipe:75%}.media.wipe img,.media.wipe object{position:absolute;inset:0}.media.wipe object{clip-path:inset(0 calc(100% - var(--wipe)) 0 0)}.media.wipe:before{content:\"\";position:absolute;z-index:3;top:0;bottom:0;left:var(--wipe);border-left:2px solid rgba(56,189,248,.9);box-shadow:0 0 0 1px rgba(2,6,23,.85)}.media.wipe:after{content:\"SVG | PNG\";position:absolute;z-index:4;left:var(--wipe);top:8px;transform:translateX(-50%);border:1px solid rgba(56,189,248,.4);border-radius:999px;background:rgba(2,6,23,.82);color:#e5e7eb;font-size:10px;font-weight:800;letter-spacing:.08em;padding:4px 8px}object,img{display:block;width:100%;height:100%;object-fit:contain;background:transparent}");
-        sb.AppendLine("@media(max-width:1200px){.pair{grid-template-columns:1fr}}@media(max-width:900px){body{padding:16px}}");
+        AppendComparisonStyle(sb);
         sb.AppendLine("</style>");
         sb.AppendLine("</head>");
         sb.AppendLine("<body>");
@@ -367,6 +361,18 @@ public static partial class GalleryWriter {
         return "<span class=\"pill " + (baseline.IsClean ? "ok" : "warn") + "\">" + baseline.ChartMatches.ToString(System.Globalization.CultureInfo.InvariantCulture) + " baseline passes</span><span class=\"pill " + (baseline.Warnings == 0 ? "ok" : "warn") + "\">" + baseline.Warnings.ToString(System.Globalization.CultureInfo.InvariantCulture) + " baseline warnings</span>";
     }
 
+    private static void AppendComparisonStyle(System.Text.StringBuilder sb) {
+        sb.AppendLine("""
+:root{color-scheme:dark;--bg:#0f172a;--panel:#111827;--frame:#020617;--line:#334155;--text:#e5e7eb;--muted:#94a3b8;--ok:#86efac;--warn:#fbbf24;--accent:#38bdf8}
+*{box-sizing:border-box}body{margin:0;padding:24px;background:var(--bg);color:var(--text);font-family:Inter,ui-sans-serif,system-ui,Segoe UI,Arial,sans-serif}
+header{max-width:1500px;margin:0 auto 24px}h1{margin:0 0 6px;font-size:24px;line-height:1.15}p{margin:0;color:var(--muted)}.summary,.family-nav{display:flex;flex-wrap:wrap;gap:8px;margin-top:14px}.pill,.family-nav a{border:1px solid var(--line);border-radius:999px;padding:6px 10px;color:#cbd5e1;font-size:12px;font-weight:700;text-decoration:none}.pill.ok,.family-nav a.ok{border-color:rgba(134,239,172,.45);color:var(--ok)}.pill.warn,.family-nav a.warn{border-color:rgba(251,191,36,.55);color:var(--warn)}a.pill:hover,.family-nav a:hover{border-color:rgba(56,189,248,.55);color:var(--accent)}main{display:grid;gap:34px;max-width:1500px;margin:0 auto}
+.family{display:grid;gap:16px}.family-head{display:flex;align-items:end;justify-content:space-between;gap:16px;border-bottom:1px solid var(--line);padding-bottom:12px}.family-title{margin:0 0 5px;font-size:19px;line-height:1.15;color:#f8fafc}.family-desc{font-size:13px}.family-count{flex:0 0 auto;border:1px solid rgba(56,189,248,.42);border-radius:999px;color:var(--accent);font-size:12px;font-weight:800;padding:6px 10px}.family-count.ok{border-color:rgba(134,239,172,.45);color:var(--ok)}.family-count.warn{border-color:rgba(251,191,36,.55);color:var(--warn)}section{background:var(--panel);border:1px solid var(--line);border-radius:8px;padding:16px;scroll-margin-top:18px}section.mismatch{border-color:rgba(251,191,36,.6)}h2{display:flex;align-items:center;justify-content:space-between;gap:12px;margin:0 0 14px;font-size:14px;color:#cbd5e1}.status{border:1px solid var(--line);border-radius:999px;padding:4px 8px;font-size:11px;font-weight:800}.status.ok{color:var(--ok);border-color:rgba(134,239,172,.45)}.status.warn{color:var(--warn);border-color:rgba(251,191,36,.55)}.pair{display:grid;grid-template-columns:repeat(2,minmax(0,1fr));gap:16px;align-items:start}section.large .pair{grid-template-columns:1fr}.wipe-figure{grid-column:1/-1;order:-1}.wipe-figure .media{min-height:360px}section.large .wipe-figure .media{min-height:min(68vh,820px)}section.large figure:not(.wipe-figure) .media{min-height:min(62vh,720px)}
+figure{margin:0;background:var(--frame);border:1px solid #1f2937;border-radius:8px;padding:10px;min-width:0}.caption{display:flex;justify-content:space-between;gap:12px;color:var(--muted);font-size:12px;margin-bottom:8px}.dims{font-variant-numeric:tabular-nums;color:#cbd5e1}.format{color:var(--accent);font-weight:800;text-decoration:none}.format:hover{text-decoration:underline}.wipe-controls{display:flex;flex-wrap:wrap;gap:6px;margin:0 0 10px}.wipe-controls input{position:absolute;opacity:0;pointer-events:none}.wipe-controls label{border:1px solid var(--line);border-radius:999px;color:#cbd5e1;cursor:pointer;font-size:11px;font-weight:800;padding:5px 9px}.wipe-controls input:focus-visible+label{outline:2px solid var(--accent);outline-offset:2px}.wipe-controls input:checked+label{background:rgba(56,189,248,.16);border-color:rgba(56,189,248,.7);color:#f8fafc}
+.media{width:100%;min-height:140px;background-color:#020617;background-image:linear-gradient(45deg,#0b1120 25%,transparent 25%),linear-gradient(-45deg,#0b1120 25%,transparent 25%),linear-gradient(45deg,transparent 75%,#0b1120 75%),linear-gradient(-45deg,transparent 75%,#0b1120 75%);background-size:24px 24px;background-position:0 0,0 12px,12px -12px,-12px 0;overflow:hidden}.media.wipe{position:relative;--wipe:50%}.wipe-controls .wipe-25:checked~.media{--wipe:25%}.wipe-controls .wipe-50:checked~.media{--wipe:50%}.wipe-controls .wipe-75:checked~.media{--wipe:75%}.media.wipe img,.media.wipe object{position:absolute;inset:0}.media.wipe object{clip-path:inset(0 calc(100% - var(--wipe)) 0 0)}.media.wipe:before{content:"";position:absolute;z-index:3;top:0;bottom:0;left:var(--wipe);border-left:2px solid rgba(56,189,248,.9);box-shadow:0 0 0 1px rgba(2,6,23,.85)}.media.wipe:after{content:"SVG | PNG";position:absolute;z-index:4;left:var(--wipe);top:8px;transform:translateX(-50%);border:1px solid rgba(56,189,248,.4);border-radius:999px;background:rgba(2,6,23,.82);color:#e5e7eb;font-size:10px;font-weight:800;letter-spacing:.08em;padding:4px 8px}object,img{display:block;width:100%;height:100%;object-fit:contain;background:transparent}
+@media(max-width:1200px){.pair{grid-template-columns:1fr}}@media(max-width:900px){body{padding:16px}}
+""");
+    }
+
     private static void AppendComparisonCard(System.Text.StringBuilder sb, ComparisonAsset pair) {
         var name = pair.Name;
         var svg = pair.SvgDimensions;
@@ -389,12 +395,27 @@ public static partial class GalleryWriter {
         sb.AppendLine("<section id=\"" + EscapeHtml(name) + "\" class=\"" + (pair.HasMatchingDimensions ? "match" : "mismatch") + largeClass + "\">");
         sb.AppendLine("<h2><span>" + EscapeHtml(name) + "</span><span class=\"status " + statusClass + "\">" + statusText + "</span></h2>");
         sb.AppendLine("<div class=\"pair\">");
-        var wipeId = Slugify(name);
-        sb.AppendLine("<figure class=\"wipe-figure\"><figcaption class=\"caption\"><span class=\"format\">WIPE</span><span class=\"dims\">" + EscapeHtml(FormatDimensions(svg)) + "</span></figcaption><div class=\"wipe-controls\" aria-label=\"" + EscapeHtml(name) + " wipe position\"><input class=\"wipe-25\" type=\"radio\" name=\"wipe-" + EscapeHtml(wipeId) + "\" id=\"wipe-" + EscapeHtml(wipeId) + "-25\"><label for=\"wipe-" + EscapeHtml(wipeId) + "-25\">SVG 25%</label><input class=\"wipe-50\" type=\"radio\" name=\"wipe-" + EscapeHtml(wipeId) + "\" id=\"wipe-" + EscapeHtml(wipeId) + "-50\" checked><label for=\"wipe-" + EscapeHtml(wipeId) + "-50\">Split</label><input class=\"wipe-75\" type=\"radio\" name=\"wipe-" + EscapeHtml(wipeId) + "\" id=\"wipe-" + EscapeHtml(wipeId) + "-75\"><label for=\"wipe-" + EscapeHtml(wipeId) + "-75\">SVG 75%</label><div class=\"media wipe\"" + ratioStyle + "><img src=\"" + EscapeHtml(name) + ".png\" alt=\"" + EscapeHtml(name) + " PNG wipe right\"><object data=\"" + EscapeHtml(name) + ".svg\" type=\"image/svg+xml\" aria-label=\"" + EscapeHtml(name) + " SVG wipe left\"></object></div></div></figure>");
-        sb.AppendLine("<figure><figcaption class=\"caption\"><a class=\"format\" href=\"" + EscapeHtml(name) + ".svg\">SVG</a><span class=\"dims\">" + EscapeHtml(svgInfo) + "</span></figcaption><div class=\"media\"" + ratioStyle + "><object data=\"" + EscapeHtml(name) + ".svg\" type=\"image/svg+xml\"></object></div></figure>");
-        sb.AppendLine("<figure><figcaption class=\"caption\"><a class=\"format\" href=\"" + EscapeHtml(name) + ".png\">PNG</a><span class=\"dims\">" + EscapeHtml(pngInfo) + "</span></figcaption><div class=\"media\"" + ratioStyle + "><img src=\"" + EscapeHtml(name) + ".png\" alt=\"" + EscapeHtml(name) + " PNG\"></div></figure>");
+        AppendComparisonWipeFigure(sb, name, svg, ratioStyle);
+        AppendComparisonAssetFigure(sb, name, "SVG", "svg", "object", svgInfo, ratioStyle);
+        AppendComparisonAssetFigure(sb, name, "PNG", "png", "img", pngInfo, ratioStyle);
         sb.AppendLine("</div>");
         sb.AppendLine("</section>");
+    }
+
+    private static void AppendComparisonWipeFigure(System.Text.StringBuilder sb, string name, AssetDimensions svg, string ratioStyle) {
+        var escapedName = EscapeHtml(name);
+        var escapedWipeId = EscapeHtml(Slugify(name));
+        sb.AppendLine("<figure class=\"wipe-figure\"><figcaption class=\"caption\"><span class=\"format\">WIPE</span><span class=\"dims\">" + EscapeHtml(FormatDimensions(svg)) + "</span></figcaption><div class=\"wipe-controls\" aria-label=\"" + escapedName + " wipe position\"><input class=\"wipe-25\" type=\"radio\" name=\"wipe-" + escapedWipeId + "\" id=\"wipe-" + escapedWipeId + "-25\"><label for=\"wipe-" + escapedWipeId + "-25\">SVG 25%</label><input class=\"wipe-50\" type=\"radio\" name=\"wipe-" + escapedWipeId + "\" id=\"wipe-" + escapedWipeId + "-50\" checked><label for=\"wipe-" + escapedWipeId + "-50\">Split</label><input class=\"wipe-75\" type=\"radio\" name=\"wipe-" + escapedWipeId + "\" id=\"wipe-" + escapedWipeId + "-75\"><label for=\"wipe-" + escapedWipeId + "-75\">SVG 75%</label><div class=\"media wipe\"" + ratioStyle + "><img src=\"" + escapedName + ".png\" alt=\"" + escapedName + " PNG wipe right\"><object data=\"" + escapedName + ".svg\" type=\"image/svg+xml\" aria-label=\"" + escapedName + " SVG wipe left\"></object></div></div></figure>");
+    }
+
+    private static void AppendComparisonAssetFigure(System.Text.StringBuilder sb, string name, string label, string extension, string mediaKind, string dimensions, string ratioStyle) {
+        var escapedName = EscapeHtml(name);
+        var escapedDimensions = EscapeHtml(dimensions);
+        var href = escapedName + "." + extension;
+        var media = string.Equals(mediaKind, "object", StringComparison.Ordinal)
+            ? "<object data=\"" + href + "\" type=\"image/svg+xml\"></object>"
+            : "<img src=\"" + href + "\" alt=\"" + escapedName + " " + label + "\">";
+        sb.AppendLine("<figure><figcaption class=\"caption\"><a class=\"format\" href=\"" + href + "\">" + label + "</a><span class=\"dims\">" + escapedDimensions + "</span></figcaption><div class=\"media\"" + ratioStyle + ">" + media + "</div></figure>");
     }
 
     private static bool IsGalleryShellFile(string fileName) =>
